requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.6b0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.8
ecdsa==0.17.0
et-xmlfile==1.1.0
fastapi==0.65.2
h11==0.12.0
idna==3.2
iniconfig==1.1.1
isort==5.9.1
kombu==5.1.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
motor==2.5.1
mypy-extensions==0.4.3
openpyxl==3.0.9
packaging==21.0
passlib==1.7.4
pathspec==0.8.1
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1